change double to long int

+++ books/cython/cp2_running_cython_code/fib.pyx
@@ -2,7 +2,7 @@
 
 def fib(int n):
     cdef int i
-    cdef double a=0.0, b=1.0
+    cdef unsigned long a=0, b=1
     for i in range(n):
         a, b = a + b, a
     return a
